+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v7.0-beta.151] - 2021-10-11
+
 * Issue **#2483** : Change reference data purge to delete entries in batchs to avoid large transactions that result in errors.
 
 * Fix performance issue with ref data range lookups.
@@ -3841,7 +3844,8 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.150...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.151...HEAD
+[v7.0-beta.151]: https://github.com/gchq/stroom/compare/v7.0-beta.150...v7.0-beta.151
 [v7.0-beta.150]: https://github.com/gchq/stroom/compare/v7.0-beta.149...v7.0-beta.150
 [v7.0-beta.149]: https://github.com/gchq/stroom/compare/v7.0-beta.148...v7.0-beta.149
 [v7.0-beta.148]: https://github.com/gchq/stroom/compare/v7.0-beta.147...v7.0-beta.148